+++ packages/server/src/util/server-import.util.ts
@@ -1,8 +1,0 @@
-export const serverImport = async <T>(path: string) => {
-  if (typeof window === "undefined") {
-    const imported: T = require(`${path}`);
-    return imported;
-  }
-
-  return undefined;
-};